@@ -48,30 +48,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.4",
+    "@angular/animations": "^7.2.0",
     "@angular/cdk": "~7.2.1",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/core": "^7.1.4",
-    "@angular/forms": "^7.1.4",
-    "@angular/http": "^7.1.4",
+    "@angular/common": "^7.2.0",
+    "@angular/compiler": "^7.2.0",
+    "@angular/core": "^7.2.0",
+    "@angular/flex-layout": "^7.0.0-beta.23",
+    "@angular/forms": "^7.2.0",
+    "@angular/http": "^7.2.0",
     "@angular/material": "^7.2.1",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
-    "core-js": "^2.6.1",
+    "@angular/platform-browser": "^7.2.0",
+    "@angular/platform-browser-dynamic": "^7.2.0",
+    "@angular/router": "^7.2.0",
+    "core-js": "^2.6.2",
     "hammerjs": "^2.0.8",
     "now": "^12.1.14",
     "rxjs": "~6.3.3",
-    "zone.js": "~0.8.26"
+    "zone.js": "^0.8.27"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.11.4",
-    "@angular/cli": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/language-service": "^7.1.4",
-    "@commitlint/cli": "^7.2.1",
-    "@commitlint/config-conventional": "^7.1.2",
+    "@angular/cli": "^7.2.1",
+    "@angular/compiler-cli": "^7.2.0",
+    "@angular/language-service": "^7.2.0",
+    "@commitlint/cli": "^7.3.1",
+    "@commitlint/config-conventional": "^7.3.1",
     "@types/jasmine": "^2.8.14",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
