@@ -1,10 +1,10 @@
-import { NgModule } from '@angular/core'
-import { MatButtonModule, MatIconModule, MatToolbarModule } from '@angular/material'
-import 'hammerjs'
+import { NgModule } from '@angular/core';
+import { MatButtonModule, MatCardModule, MatIconModule, MatToolbarModule } from '@angular/material';
+import 'hammerjs';
 
 @NgModule({
   declarations: [],
-  imports: [MatButtonModule, MatToolbarModule, MatIconModule],
-  exports: [MatButtonModule, MatToolbarModule, MatIconModule],
+  imports: [MatButtonModule, MatCardModule, MatToolbarModule, MatIconModule],
+  exports: [MatButtonModule, MatCardModule, MatToolbarModule, MatIconModule],
 })
 export class MaterialModule {}
